+++ k8s-labelstudio-vanilla/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2024-12-18
+
+* Rename variable `project` to `name`
+* Add variable `enable_ingress`
+
 # 2024-11-08
 
 * Initial implementation of vanilla setup (no E8 patches)